refactor(guard): the tests get their own file, so guard.rs stays under its ceiling
The three TOCTOU tests added by the mutation-test fix carried guard.rs to
2509 lines, nine past the 2500-line default ratchet ceiling, and the
"Check line ceiling" check went red.

Split the `#[cfg(test)] mod tests` into `guard_tests.rs` behind `#[path]`,
the shape `kernel.rs` already uses. The alternative — a `[[files]]` entry
raising the ceiling for guard.rs — buys silence rather than a smaller file,
and the same choice was made for portcullis-effects' doubles two commits
back.

No test text changed; the body is dedented by one level and nothing else.
44 guard tests pass, both clippy invocations are clean, and the ratchet
sweeps 666 files with none over the ceiling.
